Major contribution : - 3D simulation
Minor contributions :
- Deletion surface mode
- Camera rotating mode
- Solids drawing in simulation
- Particles indices
- Transparency and particles depth sorting
- Scene transparency
- Houses test scene in function
- New simple floor test Scene
- Mipmapping and anisotropic filtering in textures
- Hard coded Spot light
- New maths functions

Debugging :
- Default settings in Scene
- Centers and borders interpolation in Simulation
- Better .gitignore

+++ src/Object.hpp
@@ -68,7 +68,7 @@ class Object
 
         // Updating data on GPU
         void updateVertices(const float * const vertices, bool dynamic=false) const;
-        void updateIndices(const float * const indices, bool dynamic=false) const;
+        void updateIndices(const unsigned int * const indices, bool dynamic=false) const;
         void updateNormals(const float * const normals, bool dynamic=false) const;
         void updateUvs(const float * const uvs, bool dynamic=false) const;
         void updateColors(const float * const colors, bool dynamic=false) const;
